move Project class
