+++ fluentgen/src/main/java/com/azure/autorest/fluent/model/projectmodel/Project.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.azure.autorest.fluent.model;
+package com.azure.autorest.fluent.model.projectmodel;
 
 import com.azure.autorest.extension.base.plugin.JavaSettings;
 import com.azure.autorest.fluent.model.clientmodel.FluentClient;
